requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
requests
python-dotenv
tqdm
torch
transformers
python-jose[cryptography]
passlib[argon2]
httpx
google-cloud-vision
pdfplumber
python-multipart
pix2tex
Pillow
sympy
psycopg2-binary
sqlalchemy
alembic
authlib
